Move wordgraph display code to its own file
